+++ src/Infrastructure/IfcElementsStyleManager.js
@@ -1,6 +1,6 @@
 import IfcCustomViewSettings from './IfcCustomViewSettings'
-import {IFCPRODUCTDEFINITIONSHAPE, IFCPROPERTYSET, IFCRELDEFINESBYPROPERTIES} from 'web-ifc'
-import IfcColor from './IfcColor'
+import {IFCPRODUCTDEFINITIONSHAPE} from 'web-ifc'
+import {compileViewRules} from './ViewRulesCompiler'
 
 
 /* eslint-disable jsdoc/no-undefined-types */
@@ -18,7 +18,6 @@ export default class IfcElementsStyleManager {
     this.parser = parser
     parser._rules = rules
     parser._overrideStyles = {}
-    parser.compileViewFunction = bindCompileViewFunction(parser)
     parser.initializeLoadingState = newInitializeLoadingStateFunction(parser)
     parser.streamMesh = newStreamMeshFunction(parser)
   }
@@ -79,7 +78,8 @@ function newInitializeLoadingStateFunction(parser) {
    */
   async function initializeLoadingState(modelID) {
     if (this._rules?.length > 0) {
-      await this.compileViewFunction(this.state.api, modelID)
+      const viewSettings = await compileViewRules(this.state.api, modelID, this._rules)
+      this._overrideStyles = viewSettings
     }
     // eslint-disable-next-line new-cap
     const shapes = await this.state.api.GetLineIDsWithType(modelID, IFCPRODUCTDEFINITIONSHAPE)
@@ -89,117 +89,4 @@ function newInitializeLoadingStateFunction(parser) {
   }
 
   return initializeLoadingState.bind(parser)
-}
-
-
-/* eslint-disable new-cap */
-/* eslint-disable no-magic-numbers */
-/**
- * Returns a new compileViewFunction function that calculates the custom view
- *
- * @param {IfcParser} parser
- * @return {Function} the new compiler function
- */
-function bindCompileViewFunction(parser) {
-  /**
-   * called on initialization
-   *
-   */
-  async function compileViewFunction(api, modelID) {
-    // Apply desired logic on model before loading
-    const propRelLines = await api.GetLineIDsWithType(modelID, IFCRELDEFINESBYPROPERTIES)
-    const allPropObjects = []
-    for (let i = 0; i < propRelLines.size(); i++) {
-      allPropObjects.push( await api.GetLine(modelID, propRelLines.get(i), true))
-    }
-    const psetObjects = allPropObjects.filter((x) => x.RelatingPropertyDefinition.type === IFCPROPERTYSET)
-    // Get only sets containing PSet_vyzn.Verlust
-    const objectsAndPropVal = psetObjects.map((a) =>
-      ({o: a.RelatedObjects[0]?.expressID,
-        p: a.RelatingPropertyDefinition.HasProperties?.find(
-            (s) => s.Name.value === 'Verlust' ||
-             s.Name.value === 'SIA380-1.TransmissionHeatLoss')?.NominalValue?.value * 1})).filter((x) => x.p)
-
-    const valArr = objectsAndPropVal.map((a) => a.p)
-    const min = Math.min(valArr)
-    const max = Math.max(...valArr)
-    const entries = objectsAndPropVal.map((a) => [a.o, calculateElementColor(min, max, a)])
-    const defaultElementsColor = new IfcColor(0.96, 0.96, 0.96)
-    const viewSettings = new IfcCustomViewSettings(defaultElementsColor, Object.fromEntries(entries))
-
-    this._overrideStyles = viewSettings
-  }
-
-  return compileViewFunction.bind(parser)
-}
-
-
-/**
- * get element color based on its value
- *
- * @param {number} min
- * @param {number} max
- * @param {number} a
- * @return {IfcColor} the color
- */
-function calculateElementColor(min, max, a) {
-  const baseColor = new IfcColor(0.96, 0.96, 0.96)
-  if (a.p === 0) {
-    return baseColor
-  } else if (a.p > 0) {
-    return interpolateColors(baseColor, parseColor('#EB3324'), a.p, 0, max)
-  } else if (a.p < 0) {
-    return interpolateColors(baseColor, parseColor('#22B14C'), -1 * a.p, 0, -1 * min)
-  }
-  return baseColor
-}
-
-
-/**
- * Convert Hex Color string to IfcColor Object
- *
- * @param {string} hexColor the color in string format '#C7C7C7'
- * @return {IfcColor} the IfcColor object
- */
-function parseColor(hexColor) {
-  const parsed = hexColor.substr(1).split(/(?=(?:..)*$)/)
-      .map((a) => parseInt(a, 16) / 256)
-      .map((a) => Math.round(a * 1000) / 1000)
-  return new IfcColor(...parsed)
-}
-
-
-/**
- * get a color that lies on the scale between two colors
- *
- * @param {IfcColor} startColor
- * @param {IfcColor} targetColor
- * @param {number} value
- * @param {number} min value
- * @param {number} max value
- * @return {IfcColor} color on the scale between the start color and max color
- */
-function interpolateColors(startColor, targetColor, value, min, max) {
-  const r = changeValueScale(value, min, max, startColor.x, targetColor.x)
-  const g = changeValueScale(value, min, max, startColor.y, targetColor.y)
-  const b = changeValueScale(value, min, max, startColor.z, targetColor.z)
-  const result = new IfcColor(r, g, b)
-  return result
-}
-
-
-/**
- * convert value from one value scale to another
- *
- * @param {number} value
- * @param {number} min
- * @param {number} max
- * @param {number} targetMin
- * @param {number} targetMax
- * @return {number} the new value
- */
-function changeValueScale(value, min, max, targetMin, targetMax) {
-  const ratio = (value - min) * 1.0 / (max - min)
-  const result = targetMin + (ratio * (targetMax - targetMin))
-  return result
 }
